seperate func & now buying tickets again without restart btn is possible
