+++ index.html
@@ -5,6 +5,9 @@
     <title>🎱 행운의 로또</title>
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <link rel="stylesheet" href="./src/css/index.css" />
+    <style>
+      body{background-color:#404042};
+    </style>
   </head>
   <body>
     <div id="app" class="p-3">
